+++ leia-dropwizard/src/main/java/com/grookage/leia/dropwizard/bundle/resources/SchemaResource.java
@@ -18,13 +18,13 @@ package com.grookage.leia.dropwizard.bundle.resources;
 
 import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.Timed;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.grookage.leia.common.validation.SchemaPayloadValidator;
 import com.grookage.leia.core.exception.LeiaErrorCode;
 import com.grookage.leia.core.exception.LeiaException;
 import com.grookage.leia.core.retrieval.SchemaRetriever;
 import com.grookage.leia.models.GenericResponse;
 import com.grookage.leia.models.request.NamespaceRequest;
+import com.grookage.leia.models.request.ValidateSchemaRequest;
 import com.grookage.leia.models.schema.SchemaDetails;
 import com.grookage.leia.models.schema.SchemaKey;
 import lombok.AllArgsConstructor;
@@ -86,11 +86,11 @@ public class SchemaResource {
     @Timed
     @ExceptionMetered
     @Path("/details/validate")
-    public GenericResponse<List<String>> validateSchema(@Valid SchemaKey schemaKey,
-                                                        @Valid JsonNode jsonNode) {
-        final var schemaDetails = schemaRetriever.getSchemaDetails(schemaKey)
+    public GenericResponse<List<String>> validateSchema(@Valid final ValidateSchemaRequest validateSchemaRequest) {
+        final var schemaDetails = schemaRetriever.getSchemaDetails(validateSchemaRequest.getSchemaKey())
                 .orElseThrow(() -> LeiaException.error(LeiaErrorCode.NO_SCHEMA_FOUND));
-        final var validationErrors = SchemaPayloadValidator.validate(jsonNode, schemaDetails.getValidationType(),
+        final var validationErrors = SchemaPayloadValidator.validate(validateSchemaRequest.getJsonNode(),
+                schemaDetails.getValidationType(),
                 schemaDetails.getAttributes());
         if (validationErrors.isEmpty()) {
             return GenericResponse.<List<String>>builder()